Refactor RestaurantMenu component to utilize custom hook for menu fetching; remove redundant state management
